requirements.txt
lxml
beautifulsoup4
requests
django
furl
selenium
python-telegram-bot[ujson]